#80 [REFACTOR] MissionImageService 패키지 이전
