+++ src/main/java/com/wakeUpTogetUp/togetUp/api/mission/MissionImageService.java
@@ -1,4 +1,4 @@
-package com.wakeUpTogetUp.togetUp.api.mission.service;
+package com.wakeUpTogetUp.togetUp.api.mission;
 
 import com.wakeUpTogetUp.togetUp.api.mission.domain.CustomAnalysisEntity;
 import com.wakeUpTogetUp.togetUp.utils.file.FileUtil;
